// src/types.ts
export type NavigationPage = 
  | 'PILIH_TOKO'
  | 'DASHBOARD_UTAMA'
  | 'DASHBOARD_TOKO'
  | 'INVENTARIS'
  | 'KATEGORI'
  | 'BARANG_MASUK'
  | 'BARANG_KELUAR'
  | 'BARANG_RUSAK'
  | 'RETURN'
  | 'ABSENSI'
  | 'KARYAWAN'
  | 'LAPORAN'
  | 'ACTIVITY_LOG'
  | 'TOKO'
  | 'PENGATURAN'
  | 'PROFIL';